refactor: delete ReminderService
GSoC 2025: Review Reminders

- ReminderService was an old attempt at implementing review reminders, but the main methods inside are never actually called anymore. The only code that was still referencing the class was only accessing a constant in its companion object. Here, I've moved that constant out into IntentHandler (where it was previously used) and deleted ReminderService and its test.
- I've decided to fully delete the class instead of only marking it as ready for deprecation because the code is never actually run.
- Deleted unused strings, too.
