+++ AnkiDroid/src/main/AndroidManifest.xml
@@ -625,11 +625,6 @@
             android:exported="false"
             />
         <receiver
-            android:name=".services.ReminderService"
-            android:enabled="true"
-            android:exported="false"
-            />
-        <receiver
             android:name=".services.BootService"
             android:enabled="true"
             android:exported="false"
